# CI: Disable test-runner for Svelte 4 sandboxes