@benmccann 2025-02-06
> The timing of the flaky failures aligns with the release of [SvelteKit 2.17.0](https://github.com/sveltejs/kit/releases/tag/%40sveltejs%2Fkit%402.17.0) so that could be the reason, but I haven't confirmed this.
>
> FYI @benmccann, not something you need to worry about, but might worth knowing for you if you get other user reports about similar issues. The failures indicate that for some reason the browser sometimes navigates when/after rendering a story, which causes the Playwright tests to fail. But only for SvelteKit+Svelte 4.

FYI @dummdidumm @eltigerchino I'm case you think any of the recent changes could cause this test failure